# docs/swagger.yaml
basePath: /v1
definitions:
...
  entity.PersonalSyncSnapshot:
    properties:
      khatam_cycles:
        items:
          $ref: '#/definitions/entity.QuranKhatamCycle'
        type: array
      quran_progress:
        items:
          $ref: '#/definitions/entity.QuranReadingProgress'
        type: array
      reading_progress:
        items:
          $ref: '#/definitions/entity.ReadingProgress'
        type: array
      saved_item_ids:
        items:
          type: string
        type: array
      saved_items:
        items:
          $ref: '#/definitions/entity.SavedItem'
        type: array
      saved_items_full_resync:
        type: boolean
      server_time:
        example: "2026-06-12T03:00:00Z"
        type: string
      since:
        example: "2026-06-11T00:00:00Z"
        type: string
    type: object
...
  entity.QuranKhatamCycle:
    properties:
      completed_at:
        example: "2026-02-01T00:00:00Z"
        type: string
      completed_juz:
        items:
          type: integer
        type: array
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      juz_count:
        example: 17
        type: integer
      notes:
        example: Khatam Ramadhan
        type: string
      percent:
        example: 56.67
        type: number
      started_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  entity.QuranReadingProgress:
    properties:
      ayah_key:
        example: "73:4"
        type: string
      ayah_number:
        example: 4
        type: integer
      hizb_number:
        example: 57
        type: integer
      juz_number:
        example: 29
        type: integer
      observed_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      page_number:
        example: 574
        type: integer
      position_percent:
        example: 25
        type: number
      surah_id:
        example: 73
        type: integer
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  entity.ReadingProgress:
    properties:
      book_id:
        example: 797
        type: integer
      heading_id:
        example: 10
        type: integer
      observed_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      page_id:
        example: 12
        type: integer
      progress_percent:
        example: 32.5
        type: number
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  entity.SavedItem:
    properties:
      ayah_key:
        example: "73:4"
        type: string
      book_id:
        example: 797
        type: integer
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      from_ayah_number:
        example: 4
        type: integer
      heading_id:
        example: 10
        type: integer
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      item_type:
        example: quran_ayah
        type: string
      label:
        type: string
      note:
        type: string
      page_id:
        example: 12
        type: integer
      surah_id:
        example: 73
        type: integer
      tags:
        items:
          type: string
        type: array
      to_ayah_number:
        example: 6
        type: integer
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  v1.BatchKitabProgress:
    properties:
      book_id:
        example: 797
        minimum: 1
        type: integer
      client_observed_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      heading_id:
        example: 10
        minimum: 1
        type: integer
      page_id:
        example: 12
        minimum: 1
        type: integer
      progress_percent:
        example: 32.5
        maximum: 100
        minimum: 0
        type: number
    required:
    - book_id
    type: object
  v1.BatchKitabProgressResult:
    properties:
      error:
        example: book not found
        type: string
      progress:
        $ref: '#/definitions/entity.ReadingProgress'
      status:
        enum:
        - ok
        - error
        example: ok
        type: string
    type: object
  v1.BatchProgress:
    properties:
      kitab:
        items:
          $ref: '#/definitions/v1.BatchKitabProgress'
        maxItems: 100
        type: array
      quran:
        items:
          $ref: '#/definitions/v1.BatchQuranProgress'
        maxItems: 100
        type: array
    type: object
  v1.BatchProgressResults:
    properties:
      kitab:
        items:
          $ref: '#/definitions/v1.BatchKitabProgressResult'
        type: array
      quran:
        items:
          $ref: '#/definitions/v1.BatchQuranProgressResult'
        type: array
    type: object
  v1.BatchQuranProgress:
    properties:
      ayah_key:
        example: "73:4"
        maxLength: 16
        type: string
      client_observed_at:
        example: "2026-01-01T00:00:00Z"
        type: string
    required:
    - ayah_key
    type: object
  v1.BatchQuranProgressResult:
    properties:
      error:
        example: quran ayah not found
        type: string
      progress:
        $ref: '#/definitions/entity.QuranReadingProgress'
      status:
        enum:
        - ok
        - error
        example: ok
        type: string
    type: object
...
  v1.Error:
    properties:
      code:
        example: invalid_request_body
        type: string
      details: {}
      error:
        example: message
        type: string
      message:
        example: message
        type: string
      request_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      retry_after:
        example: 60
        type: integer
    type: object
...
  v1.SaveProgress:
    properties:
      client_observed_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      heading_id:
        example: 10
        minimum: 1
        type: integer
      page_id:
        example: 12
        minimum: 1
        type: integer
      progress_percent:
        example: 32.5
        maximum: 100
        minimum: 0
        type: number
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Surau classical book reader API
  title: Go Clean Template API
  version: "1.0"
paths:
...
  /me/progress/{book_id}:
    get:
      description: Return the authenticated user's reading position for one book.
      operationId: get-progress
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ReadingProgress'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Get reading progress
      tags:
      - me
    put:
      consumes:
      - application/json
      description: Upsert the authenticated user's reading position for one book.
        Older client_observed_at events do not roll back progress.
      operationId: save-progress
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: integer
      - description: Reading position
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.SaveProgress'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.ReadingProgress'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Save reading progress
      tags:
      - me
...
  /me/progress/batch:
    post:
      consumes:
      - application/json
      description: Replay an offline autosave queue in one request (max 100 kitab
        and 100 Quran entries). Entries are processed in order and reported one-to-one
        in the response; stale client_observed_at entries are accepted but never roll
        progress back. Domain failures (e.g. deleted book) mark only that entry as
        error.
      operationId: batch-save-progress
      parameters:
      - description: Offline progress queue
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.BatchProgress'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.BatchProgressResults'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Batch save reading progress
      tags:
      - me
...
  /me/sync:
    get:
      description: Return reading progress (kitab and Quran), saved items, and khatam
        cycles changed at or after the since cursor; omit since for a full snapshot.
        Delivery is at-least-once (a server-side overlap window re-sends recent rows),
        so clients must upsert idempotently by key and store server_time as the next
        cursor. saved_item_ids lists every current saved-item ID for delete reconciliation;
        when saved_items_full_resync=true the ID list was over the server cap (10000)
        and arrives empty — rebuild the local saved-items store by paging GET /me/saved-items
        instead of deleting by absence.
      operationId: sync-personal-data
      parameters:
      - description: RFC3339 cursor from the previous response's server_time
        example: "2026-06-11T00:00:00Z"
        in: query
        name: since
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.PersonalSyncSnapshot'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Sync personal reader data
      tags:
      - me
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"